Extract into new VaultTimeoutSettingsService
