+++ libs/common/src/abstractions/vaultTimeout/vaultTimeout.service.ts
@@ -2,9 +2,4 @@ export abstract class VaultTimeoutService {
   checkVaultTimeout: () => Promise<void>;
   lock: (userId?: string) => Promise<void>;
   logOut: (userId?: string) => Promise<void>;
-  setVaultTimeoutOptions: (vaultTimeout: number, vaultTimeoutAction: string) => Promise<void>;
-  getVaultTimeout: () => Promise<number>;
-  isPinLockSet: () => Promise<[boolean, boolean]>;
-  isBiometricLockSet: () => Promise<boolean>;
-  clear: (userId?: string) => Promise<any>;
 }
