+++ libs/common/src/abstractions/vaultTimeout/vaultTimeoutSettings.service.ts
@@ -1,0 +1,7 @@
+export abstract class VaultTimeoutSettingsService {
+  setVaultTimeoutOptions: (vaultTimeout: number, vaultTimeoutAction: string) => Promise<void>;
+  getVaultTimeout: (userId?: string) => Promise<number>;
+  isPinLockSet: () => Promise<[boolean, boolean]>;
+  isBiometricLockSet: () => Promise<boolean>;
+  clear: (userId?: string) => Promise<any>;
+}
